feat: cmd组织结构调整
