--- src/cmd/config/index.ts
@@ -1,1 +1,0 @@
-export { options as ConfigPlugin } from "./options";
