--- src/cmd/dev/index.ts
@@ -1,1 +1,0 @@
-export { options as DevPlugin } from "./options";
