--- src/cmd/middleware/index.ts
@@ -1,1 +1,0 @@
-export { options as MiddlewarePlugin } from "./options";
