+++ src/cmd/process/check.ts
