+++ src/cmd/process/compile-configs.ts
@@ -2,7 +2,6 @@ import { preConfigCompiler } from "../cmd";
 
 const { CONFIG_ROOT, OUTPUT_ROOT, FORCE, ENABLED, __TSCONFIG } = process.env;
 
-// @ts-ignore
 try {
   const results = preConfigCompiler({
     tsconfig: __TSCONFIG === "-" ? undefined : __TSCONFIG,
