+++ src/cmd/process/middleware-run.ts
@@ -1,10 +1,9 @@
-import { middlewareCompileFn } from "../builders/middleware-cmp";
+import { preMiddlewareCompiler } from "../cmd";
 
 const { FOLDER_ROOT, OUTPUT_ROOT, FORCE, ENABLED, __TSCONFIG } = process.env;
 
-// @ts-ignore
 try {
-  const results = middlewareCompileFn({
+  const results = preMiddlewareCompiler({
     tsconfig: __TSCONFIG === "-" ? undefined : __TSCONFIG,
     rootFolder: FOLDER_ROOT === "-" ? undefined : FOLDER_ROOT,
     outFolder: OUTPUT_ROOT === "-" ? undefined : OUTPUT_ROOT,
