--- src/cmd/router/index.ts
@@ -1,1 +1,0 @@
-export { options as RouterPlugin } from "./options";
